Move auth and notifications actions into sagas
