--- assets/js/state/auth.js
@@ -1,6 +1,0 @@
-export const PERFORM_LOGIN = 'PERFORM_LOGIN';
-
-export const performLogin = ({ username, password }) => ({
-  type: PERFORM_LOGIN,
-  payload: { username, password },
-});
