+++ assets/js/state/sagas/databases.test.js
@@ -29,7 +29,7 @@ import {
   databaseInstanceFactory,
 } from '@lib/test-utils/factories';
 import { networkClient } from '@lib/network';
-import { notify } from '@state/notifications';
+import { notify } from '@state/sagas/notifications';
 
 const axiosMock = new MockAdapter(networkClient);
