+++ assets/js/state/sagas/notifications.jsx
@@ -2,7 +2,13 @@
 import React from 'react';
 import { takeEvery } from 'redux-saga/effects';
 import { toast } from 'react-hot-toast';
-import { NOTIFICATION } from '@state/notifications';
+
+export const NOTIFICATION = 'NOTIFICATION';
+
+export const notify = ({ text, icon }) => ({
+  type: NOTIFICATION,
+  payload: { text, icon },
+});
 
 export function* notification({ payload }) {
   const { text, icon } = payload;
